refactor api_helpers; handle exceptions better; use oauth_required for scoping

+++ app/controllers/api/contact_assignments_controller.rb
@@ -1,41 +1,32 @@
 class Api::ContactAssignmentsController < ApiController
   require 'api_helper'
   include ApiHelper
-  oauth_required
-  before_filter :get_organization
-  skip_before_filter :verify_authenticity_token
-  skip_before_filter :authenticate_user!
-
+  oauth_required :scope => "contact_assignment"
+  before_filter :valid_request_before, :organization_allowed?, :authorized_leader?
+  skip_before_filter :verify_authenticity_token, :authenticate_user!
+  rescue_from Exception, :with => :render_json_error
+  
   def create_1
-    json_render = {:error => "You did not provide the appropriate parameters to create a contact assignment."}
-    if ((params[:org_id].present? || params[:org].present?) && params[:id].present? && !@organization.empty? && params[:assign_to].present?)
+    raise ContactAssignmentCreateParamsError unless ((params[:org_id].present? || params[:org].present?) && params[:id].present? && params[:assign_to].present?)
+    if !@organization.empty?
       ids = params[:id].split(',')
       ContactAssignment.where(:person_id => ids, :organization_id => @organization.id).destroy_all
-      if params[:assign_to].present?
-        @assign_to = Person.find(params[:assign_to])
-        params[:ids].each do |id|
-          ContactAssignment.create!(:person_id => id, :organization_id => @organization.id, :assigned_to_id => @assign_to.id)
-          json_render = []
-        end
+      @assign_to = Person.find(params[:assign_to])
+      params[:ids].each do |id|
+        ContactAssignment.create!(:person_id => id, :organization_id => @organization.id, :assigned_to_id => @assign_to.id)
       end
     end
-    render :json => JSON::pretty_generate(json_render)
+    render :json => JSON::pretty_generate('[]')
   end
   
   def destroy_1
-    valid_fields = valid_request_with_rescue(request)
-    return render :json => valid_fields if valid_fields.is_a? Hash
-    return render :json => {:error => "You do not have the appropriate organization memberships to view this data."} unless organization_allowed?
-    json_render = {:error => "You did not provide the appropriate parameters to delete a contact assignment."}
-    
+    raise ContactAssignmentDeleteParamsError unless ((params[:org_id].present? || params[:org].present?) && params[:id].present?)
     ids = params[:id].split(',')
-    
-    if ((params[:org_id].present? || params[:org].present?) && params[:id].present? && !@organization.empty?)
-      ContactAssignment.where(:person_id => ids, :organization_id => @organization.first.id).destroy_all
-      json_render = []
-    end
-    render :json => JSON::pretty_generate(json_render)
+    ContactAssignment.where(:person_id => ids, :organization_id => @organization.first.id).destroy_all unless @organization.empty?
+    render :json => JSON::pretty_generate('[]')
   end
+  
+  private
   
   def get_organization
     org = params[:org_id].present? ? params[:org_id] : params[:org]
